feat: update structure project
