+++ crud-api/.prettierignore
@@ -1,3 +1,4 @@
 node_modules
 .prettierignore
 .eslintignore
+dist
