+++ crud-api/nodemon.json
@@ -2,5 +2,5 @@
     "watch": ["src"],
     "ext": "ts,json",
     "ignore": ["src/**/*.test.ts"],
-    "exec": "tsx ./src/main.ts"
+    "exec": "ts-node ./src/main.ts"
 }
