+++ crud-api/tsconfig.json
@@ -1,8 +1,8 @@
 {
     "compilerOptions": {
         "target": "ESNext",
-        "module": "ESNext",
-        "moduleResolution": "node",
+        "module": "CommonJS",
+        "moduleResolution": "Node",
         "esModuleInterop": true,
         "strict": true,
         "skipLibCheck": true,
@@ -13,5 +13,5 @@
         "resolveJsonModule": true
     },
     "include": ["src/**/*.ts"],
-    "exclude": ["node_modules"]
+    "exclude": ["node_modules", "**/*.spec.ts"]
 }
